requirements.txt
requests
beautifulsoup4
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv